fix:调整数据算法

+++ src/main/java/com/tangerineSpecter/oms/system/domain/entity/DataTradeRecord.java
@@ -38,7 +38,7 @@ public class DataTradeRecord {
     /**
      * 收益值
      */
-    private Integer incomeValue;
+    private BigDecimal incomeValue;
     /**
      * 收益率
      */
